# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Token Vault API specification
 * OpenAPI spec version: 0.2.0
 */
import type { ActivityItemType } from "./activityItemType";

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  vaultName: string;
  tokenSymbol: string;
  amount: string;
  timestamp: Date;
}