# lib/api-zod/src/generated/types/createLiquidityLockBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Token Vault API specification
 * OpenAPI spec version: 0.2.0
 */

export interface CreateLiquidityLockBody {
  name: string;
  tokenPair: string;
  dex: string;
  chain: string;
  lpTokenAmount: string;
  /** @minimum 1 */
  lockDays: number;
  contractAddress?: string;
}